refactor: InputView 관련 함수들을 InputView 하나의 모듈에서 관리하도록 변경
