+++ src/utils/index.js
@@ -1,4 +1,3 @@
-export * from './view.js';
 export * from './validate.js';
 export * from './racingTrack.js';
 export * from './racingWinners.js';
